feat(gpt): generatSummary for articles
